Bump IconExplorer to RN 0.61.5
