+++ Examples/IconExplorer/package.json
@@ -3,25 +3,28 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
+    "android": "react-native run-android",
+    "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest"
   },
   "dependencies": {
     "ramda": "^0.26.1",
-    "react": "16.8.6",
-    "react-native": "0.60.5",
-    "react-native-gesture-handler": "^1.3.0",
-    "react-native-screens": "^1.0.0-alpha.23",
+    "react": "16.9.0",
+    "react-native": "0.61.5",
+    "react-native-gesture-handler": "^1.5.1",
+    "react-native-screens": "^2.0.0-alpha.12",
     "react-native-vector-icons": "*",
-    "react-navigation": "^3.11.1"
+    "react-navigation": "^4.0.10",
+    "react-navigation-stack": "^1.10.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/runtime": "^7.5.5",
+    "@babel/core": "^7.7.4",
+    "@babel/runtime": "^7.7.4",
     "babel-jest": "^24.9.0",
     "jest": "^24.9.0",
-    "metro-react-native-babel-preset": "^0.56.0",
-    "react-test-renderer": "16.8.6"
+    "metro-react-native-babel-preset": "^0.57.0",
+    "react-test-renderer": "16.9.0"
   },
   "jest": {
     "preset": "react-native"
