+++ Examples/IconExplorer/src/IconList.js
