+++ Examples/IconExplorer/src/icon-sets.js
