+++ Examples/IconExplorer/src/react-native.android.js
