+++ Examples/IconExplorer/src/react-native.ios.js
