+++ Examples/IconExplorer/src/react-native.osx.js
