+++ Examples/IconExplorer/src/react-native.windows.js
